Figure out how to generate lcov files
The build was finally passing when trying to use the coveralls_reborn
gem, but it still wasn't posting to the API correctly. It would just get
a swallowed error like

  [Coveralls] Submitting to https://coveralls.io/api/v1
  [Coveralls] Couldn't find a repository matching this job.

Poking around the API and the source code, I wonder if it's just a
fool's errand to try to get the gem to recognize GitHub Actions. Too
many environment variables to set that give coveralls all this context.

Okay, so let's try moving back to coverallsapp/github-action@v1.1.2,
which means I need lcov files. Turns out there's a plugin for simplecov
that ostensibly generates such output. So I'm going to try pointing the
action over to that file and see if it can figure out how to get the
results from point A to point B.