Relicense under MIT
IANAL, but MIT just seems simpler, and I'm kind of getting anxiety
reading through the Apache text. :-/ I gave credit to the original
newrelic_rpm work in the README, so maybe that's enough to comply with
their licensing terms? Even though (a) I've significantly rewritten some
parts of those regular expressions and (b) if they wanna come after me
for licensing issues on a fuckin regular expression, so be it. Anyway,
it probably won't matter.